+++ tests/test_extrusion_ladder_graph.py
@@ -30,68 +30,12 @@ from pychoreo_examples.extrusion.transition_planner import solve_transition_betw
 
 import ikfast_kuka_kr6_r900
 
-from compas.robots import RobotModel
-from compas_fab.robots import Robot as RobotClass
-from compas_fab.robots import RobotSemantics
-
-@pytest.fixture
-def problem():
-    return 'four-frame'
-    # return 'simple_frame'
-
-def get_problem_path(problem):
-    EXTRUSION_DIRECTORY = pychoreo_examples.get_data('assembly_instances/extrusion')
-    EXTRUSION_FILENAMES = {
-        'topopt-100': 'topopt-100_S1_03-14-2019_w_layer.json',
-        'voronoi': 'voronoi_S1_03-14-2019_w_layer.json',
-        'four-frame': 'four-frame.json',
-        'simple_frame': 'simple_frame.json',
-    }
-    EXTRUSION_SEQ_FILENAMES = {
-        'four-frame': 'four-frame_solution_regression-z.json',
-        'simple_frame': 'simple_frame_solution_regression-z.json',
-    }
-    here = os.path.dirname(__file__)
-    assert problem in EXTRUSION_FILENAMES and problem in EXTRUSION_SEQ_FILENAMES
-    return os.path.join(EXTRUSION_DIRECTORY, EXTRUSION_FILENAMES[problem]), \
-           os.path.join(here, 'test_data', EXTRUSION_SEQ_FILENAMES[problem])
-
-def get_robot_data():
-    URDF_PATH = 'models/kuka_kr6_r900/urdf/kuka_kr6_r900_extrusion.urdf'
-    robot_urdf = pychoreo_examples.get_data(URDF_PATH)
-    SRDF_PATH = 'models/kuka_kr6_r900/srdf/kuka_kr6_r900_extrusion.srdf'
-    robot_srdf = pychoreo_examples.get_data(SRDF_PATH)
-
-    WORKSPACE_URDF_PATH = 'models/kuka_kr6_r900/urdf/mit_3-412_workspace.urdf'
-    workspace_urdf = pychoreo_examples.get_data(WORKSPACE_URDF_PATH)
-    WORKSPACE_SRDF_PATH = 'models/kuka_kr6_r900/srdf/mit_3-412_workspace.srdf'
-    workspace_srdf = pychoreo_examples.get_data(WORKSPACE_SRDF_PATH)
-
-    move_group = 'manipulator_ee'
-
-    robot_model = RobotModel.from_urdf_file(robot_urdf)
-    robot_semantics = RobotSemantics.from_srdf_file(robot_srdf, robot_model)
-    robot = RobotClass(robot_model, semantics=robot_semantics)
-
-    base_link_name = robot.get_base_link_name(group=move_group)
-    ee_link_name = robot.get_end_effector_link_name(group=move_group)
-    ik_joint_names = robot.get_configurable_joint_names(group=move_group)
-    disabled_self_collision_link_names = robot_semantics.get_disabled_collisions()
-    tool_root_link_name = 'eef_base_link' # TODO: should be derived from SRDF as well
-
-    workspace_model = RobotModel.from_urdf_file(workspace_urdf)
-    workspace_semantics = RobotSemantics.from_srdf_file(workspace_srdf, workspace_model)
-    workspace_robot_disabled_link_names = workspace_semantics.get_disabled_collisions()
-
-    return (robot_urdf, base_link_name, tool_root_link_name, ee_link_name, ik_joint_names, disabled_self_collision_link_names), \
-           (workspace_urdf, workspace_robot_disabled_link_names)
-
-def load_extrusion_end_effector():
+def load_extrusion_end_effector(ee_urdf_path):
     with HideOutput():
-        ee = load_pybullet(pychoreo_examples.get_data('models/kuka_kr6_r900/urdf/extrusion_end_effector.urdf'))
+        ee = load_pybullet(ee_urdf_path)
     return ee
 
-def build_extrusion_cartesian_process(elements, node_points, robot, ik_fn, ik_joint_names, base_link_name, tool_from_root=None, viz_step=False):
+def build_extrusion_cartesian_process(elements, node_points, robot, ik_fn, ik_joint_names, base_link_name, extrusion_end_effector, tool_from_root=None, viz_step=False):
     def get_sample_ik_fn(robot, ik_fn, ik_joint_names, base_link_name, tool_from_root=None):
         def sample_ik_fn(world_from_tcp):
             if tool_from_root:
@@ -103,7 +47,7 @@ def build_extrusion_cartesian_process(elements, node_points, robot, ik_fn, ik_jo
     sample_ik_fn = get_sample_ik_fn(robot, ik_fn, ik_joint_names, base_link_name, tool_from_root)
 
     # load EE body, for debugging purpose
-    ee_body = load_extrusion_end_effector()
+    ee_body = load_extrusion_end_effector(extrusion_end_effector)
     ik_joints = joints_from_names(robot, ik_joint_names)
 
     cart_traj_dict = {}
@@ -156,10 +100,10 @@ def build_extrusion_cartesian_process(elements, node_points, robot, ik_fn, ik_jo
     return cart_traj_dict
 
 @pytest.mark.extrusion
-def test_extrusion_ladder_graph(problem, viewer):
+def test_extrusion_ladder_graph(viewer, extrusion_problem_path, extrusion_robot_data, extrusion_end_effector):
     # * create robot and pb environment
     (robot_urdf, base_link_name, tool_root_link_name, ee_link_name, ik_joint_names, disabled_self_collision_link_names), \
-        (workspace_urdf, workspace_robot_disabled_link_names) = get_robot_data()
+        (workspace_urdf, workspace_robot_disabled_link_names) = extrusion_robot_data
     connect(use_gui=viewer)
     with HideOutput():
         robot = load_pybullet(robot_urdf, fixed_base=True)
@@ -181,7 +125,7 @@ def test_extrusion_ladder_graph(problem, viewer):
     tool_from_root = get_relative_pose(robot, root_link, tool_link)
 
     # * get problem & pre-computed json file paths
-    file_path, seq_file_path = get_problem_path(problem)
+    file_path, seq_file_path = extrusion_problem_path
 
     # * load shape (get nodal positions)
     elements, node_points, ground_nodes = load_extrusion(file_path)
@@ -199,7 +143,7 @@ def test_extrusion_ladder_graph(problem, viewer):
 
     # * create cartesian processes without a sequence being given, with random pose generators
     cart_process_dict = build_extrusion_cartesian_process(elements, node_points, robot, ik_fn, ik_joint_names,
-        base_link_name, tool_from_root, viz_step=False)
+        base_link_name, extrusion_end_effector, tool_from_root, viz_step=False)
 
     # * load precomputed sequence
     with open(seq_file_path, 'r') as f:
@@ -232,7 +176,7 @@ def test_extrusion_ladder_graph(problem, viewer):
         return ee_pose_map_fn
 
     with WorldSaver():
-        ee_body = load_extrusion_end_effector()
+        ee_body = load_extrusion_end_effector(extrusion_end_effector)
         ee_pose_map_fn = get_ee_pose_map_fn(roll_disc, pitch_disc)
 
         # * building collision function based on the given sequence
@@ -301,6 +245,7 @@ def test_extrusion_ladder_graph(problem, viewer):
                                                                    obstacles=[], return2idle=return2idle)
     assert all(isinstance(tt, MotionTrajectory) for tt in transition_traj)
     if return2idle:
+        transition_traj[-1].tag = 'return2idle'
         assert len(transition_traj)-1 == len(print_trajs)
     else:
         assert len(transition_traj) == len(print_trajs)
@@ -324,14 +269,15 @@ def test_extrusion_ladder_graph(problem, viewer):
         display_trajectories(robot_urdf, ik_joint_names, ee_link_name, node_points, ground_nodes, full_trajs,
                              workspace_urdf=workspace_urdf, animate=True, cart_time_step=0.02, tr_time_step=0.05)
 
+
 @pytest.mark.extrusion_viz
-def test_parse_and_visualize_results(problem, viewer):
+def test_parse_and_visualize_results(viewer, extrusion_problem_path, extrusion_robot_data, extrusion_end_effector):
     # * create robot and pb environment
     (robot_urdf, base_link_name, tool_root_link_name, ee_link_name, ik_joint_names, disabled_self_collision_link_names), \
-        (workspace_urdf, workspace_robot_disabled_link_names) = get_robot_data()
+        (workspace_urdf, workspace_robot_disabled_link_names) = extrusion_robot_data
 
     # * get problem & pre-computed json file paths
-    file_path, _ = get_problem_path(problem)
+    file_path, _ = extrusion_problem_path
 
     # * load shape (get nodal positions)
     elements, node_points, ground_nodes = load_extrusion(file_path)
@@ -361,4 +307,4 @@ def test_parse_and_visualize_results(problem, viewer):
     # visualize plan
     if viewer:
         display_trajectories(robot_urdf, ik_joint_names, ee_link_name, node_points, ground_nodes, full_trajs,
-                             workspace_urdf=workspace_urdf, animate=True, cart_time_step=0.06, tr_time_step=0.05)
+                             workspace_urdf=workspace_urdf, animate=True, cart_time_step=0.1, tr_time_step=0.05)